requirements.txt
langchain==0.2.6
pytesseract==0.3.10
pdf2image==1.17.0
pillow==10.4.0
langchain-community==0.2.6
transformers==4.42.3
sentence-transformers==3.0.1
chromadb==0.5.3
accelerate==0.32.1
bitsandbytes==0.43.1
sentencepiece==0.2.0
peft==0.11.1
streamlit==1.36.0
cohere==5.5.8
langchain-google-genai==1.0.7
